+++ app/routes/components/Flex/changelog.codegen.ts
@@ -3,4 +3,6 @@
 
 export const changelog = ''
   + '- 2025-10-02: Initial version.\n'
+  + '- 2025-10-04: Support custom `style` prop and spread additional props. Extend\n'
+  + '  from `ComponentProps<"div">` for better TypeScript support.\n'
   ;