+++ src/components/Flex/index.ts
@@ -1,3 +1,3 @@
-// Flex from copy-ui @ 2025-10-02.
+// Flex from copy-ui @ 2025-10-04.
 
 export { Flex } from "./Flex";