+++ src/components/Flex/CHANGELOG.md
@@ -1,1 +1,3 @@
 - 2025-10-02: Initial version.
+- 2025-10-04: Support custom `style` prop and spread additional props. Extend
+  from `ComponentProps<"div">` for better TypeScript support.